feat: useColorPalette 훅 추가 및 닉네임 배경 설정 기능 추가
